requirements.txt
Django==3.2.8
python-dateutil==2.8.2
google-api-python-client==2.26.1
google-auth-oauthlib==0.4.6
gspread==4.0.1
fuzzywuzzy==0.18.0
num2word==1.0.1
numpy==1.26.4
pandas==1.5.3
psycopg2-binary==2.9.5
python-Levenshtein==0.12.2
gunicorn==20.1.0
pycodestyle==2.8.0
redis==3.5.3
celery==5.2.7
newrelic==7.2.1.168
git+https://git@github.com/quizitive/django-bulk-update-or-create.git
django-redis==5.0.0
django-redis-sessions==0.6.2
django-debug-toolbar==3.2.2
pytz==2021.3
sentry-sdk==1.4.3
django-sslserver==0.22
fontawesome-free==5.15.4
social-auth-app-django==5.0.0
sendgrid==6.8.2
django-ckeditor==6.1.0
Pillow==8.3.2
django-object-actions==3.0.2
channels==3.0.4
channels-redis==3.3.1
django-sortedm2m-filter-horizontal-widget==2.0.1
openpyxl==3.0.9
pydot==1.4.2
emoji==1.6.1
git+https://git@github.com/quizitive/django-request.git
qrcode==7.3.1
xhtml2pdf==0.2.7
black==24.10.0
pre-commit==4.0.1
